+++ db/migrate/20220701040939_create_stock_transactions.rb
@@ -5,7 +5,7 @@ class CreateStockTransactions < ActiveRecord::Migration[6.1]
       t.string  :symbol
       t.string  :company_name
       t.decimal :unit_price
-      t.decimal :quantity
+      t.float   :quantity
       t.decimal :amount
       t.integer :transaction_type
